+++ Items/AgletOfTheWind/AgletOfTheWind.cs
@@ -1,0 +1,47 @@
+﻿using Terraria.ID;
+using Terraria;
+using Terraria.ModLoader;
+using System.Collections.Generic;
+
+namespace Combinations.Items.AgletOfTheWind
+{
+    public class AgletOfTheWind : CombinationsBaseModItem
+    {
+        public override void SetStaticDefaults()
+        {
+            Tooltip.SetDefault("12% increased movement speed");
+        }
+
+        public override void SetDefaults()
+        {
+            Item.accessory = true;
+            Item.width = 22;
+            Item.height = 22;
+            Item.value = Item.sellPrice(0, 1, 50);
+            Item.rare = ItemRarityID.Orange;
+            Item.stack = 1;
+        }
+
+        public override void AddRecipes()
+        {
+            Recipe recipe = CreateRecipe();
+            recipe.AddIngredient(ItemID.Aglet);
+            recipe.AddIngredient(ItemID.AnkletoftheWind);
+            recipe.AddTile(TileID.TinkerersWorkbench);
+            recipe.Register();
+        }
+
+        public override void UpdateAccessory(Player player, bool hideVisual)
+        {
+            player.moveSpeed += 0.12f;
+        }
+
+        public static int ItemType() => ModContent.ItemType<AgletOfTheWind>();
+
+        public override List<int> IncompatibleAccessories() =>
+            new List<int>()
+            {
+                ItemType(),
+            };
+    }
+}